requirements.txt
openai>=1.0.0
numpy
faiss-cpu>=1.8.0
python-dotenv>=1.0.0